requirements.txt
pandas==1.5.1
Pillow==9.4.0
requests==2.28.1
streamlit==1.14.0
